+++ tests/testbed/003-receiver-name-resolves/app.rb
@@ -1,0 +1,13 @@
+class Widget
+  def ship_it
+  end
+end
+class Job
+  def initialize(widget)
+    @widget = widget
+  end
+
+  def run
+    @widget.ship_it
+  end
+end