+++ tests/testbed/004-receiver-name-is-ambiguous/app.rb
@@ -1,0 +1,17 @@
+class Alpha
+  def ship
+  end
+end
+class Widget
+  def ship
+  end
+end
+class Zeta
+  def ship
+  end
+end
+class Job
+  def run
+    @widget.ship
+  end
+end