+++ tests/testbed/005-competing-definition-blocks-promotion/app.rb
@@ -1,0 +1,12 @@
+class Widget
+  def ship
+  end
+end
+class Job
+  def ship
+  end
+
+  def run
+    @widget.ship
+  end
+end