+++ tests/testbed/009-finder-types-the-local/app.rb
@@ -1,0 +1,10 @@
+class Widget
+  def ship
+  end
+end
+class Job
+  def run
+    w = Widget.find(1)
+    w.ship
+  end
+end